+++ proj/src/kernel/pipe_file.cpp
@@ -28,7 +28,7 @@ Pipe_Out_File::Pipe_Out_File(std::shared_ptr<Pipe> p) : pipe(std::move(p)) {}
 bool Pipe_Out_File::read(size_t size, char *out_buffer, size_t &written) {
     // write contents of vector<char> to the given buffer
     auto data = pipe->Read(size / sizeof(char));
-    for (int i = 0; i < data.size(); i++) {
+    for (size_t i = 0; i < data.size(); i++) {
         out_buffer[i] = data.at(i);
     }
     written = data.size() * sizeof(char);
